pyproject.toml
[tool.poetry]
name = "webcap"
version = "0.1.0"
description = ""
authors = ["TheTechromancer"]
license = "GPL-v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
httpx = "^0.28.0"
websockets = "^14.1"
orjson = "^3.10.12"
extractous = "^0.2.0"
uvloop = "^0.21.0"
imagehash = "^4.3.1"

[tool.poetry.scripts]
webcap = "webcap.cli:main"

[tool.ruff]
line-length = 119
lint.ignore = ["E402", "E721", "E741", "F401", "F403", "F405", "E713"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"
pytest-httpserver = "^1.1.0"
pytest-asyncio = "^0.24.0"
poetry-dynamic-versioning = "^1.4.1"
pytest-cov = "^6.0.0"
lxml = "^5.3.0"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
metadata = true
format = "v0.1.{distance}"
